ENH: Many fixes and upgrades (#47)
- Fixes #44 
- Fixes #43
- Fixes #40
- Fixes #39
- Fixes #30
- Fixes #29 

Incorporates numerous other minor enhancements and fixes.

+++ q2_feature_classifier/_taxonomic_classifier.py
@@ -10,7 +10,9 @@ import json
 import tarfile
 import os
 
+import sklearn
 from sklearn.externals import joblib
+from sklearn.pipeline import Pipeline
 import qiime2.plugin
 import qiime2.plugin.model as model
 
@@ -44,55 +46,79 @@ class TaxonomicClassifierDirFmt(model.DirectoryFormat):
     sklearn_pipeline = model.File('sklearn_pipeline.tar', format=PickleFormat)
 
 
+class TaxonomicClassiferTemporaryPickleDirFmt(model.DirectoryFormat):
+    version_info = model.File('sklearn_version.json', format=JSONFormat)
+    sklearn_pipeline = model.File('sklearn_pipeline.tar', format=PickleFormat)
+
+
 # Transformers
 @plugin.register_transformer
-def _1(dirfmt: TaxonomicClassifierDirFmt) -> dict:
-    # Note: the next two lines will likely disappear when cleanup of the
-    # views/transformers API takes place.
-    preprocess_params = dirfmt.preprocess_params.view(JSONFormat)
+def _1(dirfmt: TaxonomicClassiferTemporaryPickleDirFmt) -> Pipeline:
+    sklearn_version = dirfmt.version_info.view(dict)['sklearn-version']
+    if sklearn_version != sklearn.__version__:
+        raise ValueError('The scikit-learn version (%s) used to generate this'
+                         ' artifact does not match the current version'
+                         ' of scikit-learn installed (%s). Please retrain your'
+                         ' classifier for your current deployment to prevent'
+                         ' data-corruption errors.'
+                         % (sklearn_version, sklearn.__version__))
+
     sklearn_pipeline = dirfmt.sklearn_pipeline.view(PickleFormat)
 
-    with preprocess_params.open() as fh:
-        params = json.load(fh)
-
     with tarfile.open(str(sklearn_pipeline)) as tar:
-        dirname = os.path.dirname(str(sklearn_pipeline))
+        tmpdir = model.DirectoryFormat()
+        dirname = str(tmpdir)
         tar.extractall(dirname)
         pipeline = joblib.load(os.path.join(dirname, 'sklearn_pipeline.pkl'))
         for fn in tar.getnames():
             os.unlink(os.path.join(dirname, fn))
 
-    return {'params': params, 'pipeline': pipeline}
+    return pipeline
 
 
 @plugin.register_transformer
-def _2(data: dict) -> TaxonomicClassifierDirFmt:
-    if 'params' not in data:
-        raise ValueError('classifier does not contain params')
-    if 'pipeline' not in data:
-        raise ValueError('classifier does not contain pipeline')
-
-    preprocess_params = JSONFormat()
-    with preprocess_params.open() as fh:
-        json.dump(data['params'], fh)
-
+def _2(data: Pipeline) -> TaxonomicClassiferTemporaryPickleDirFmt:
     sklearn_pipeline = PickleFormat()
     with tarfile.open(str(sklearn_pipeline), 'w') as tar:
-        pf = os.path.join(os.path.dirname(str(sklearn_pipeline)),
-                          'sklearn_pipeline.pkl')
-        for fn in joblib.dump(data['pipeline'], pf):
+        tmpdir = model.DirectoryFormat()
+        pf = os.path.join(str(tmpdir), 'sklearn_pipeline.pkl')
+        for fn in joblib.dump(data, pf):
             tar.add(fn, os.path.basename(fn))
             os.unlink(fn)
 
-    dirfmt = TaxonomicClassifierDirFmt()
-    dirfmt.preprocess_params.write_data(preprocess_params, JSONFormat)
+    dirfmt = TaxonomicClassiferTemporaryPickleDirFmt()
+    dirfmt.version_info.write_data(
+        {'sklearn-version': sklearn.__version__}, dict)
     dirfmt.sklearn_pipeline.write_data(sklearn_pipeline, PickleFormat)
 
     return dirfmt
 
 
+@plugin.register_transformer
+def _3(dirfmt: TaxonomicClassifierDirFmt) -> dict:
+    raise ValueError('The scikit-learn version could not be determined for'
+                     ' this artifact, please retrain your classifier for your'
+                     ' current deployment to prevent data-corruption errors.')
+
+
+@plugin.register_transformer
+def _4(fmt: JSONFormat) -> dict:
+    with fmt.open() as fh:
+        return json.load(fh)
+
+
+@plugin.register_transformer
+def _5(data: dict) -> JSONFormat:
+    result = JSONFormat()
+    with result.open() as fh:
+        json.dump(data, fh)
+    return result
+
+
 # Registrations
 plugin.register_semantic_types(TaxonomicClassifier)
-plugin.register_formats(PickleFormat, JSONFormat, TaxonomicClassifierDirFmt)
+plugin.register_formats(PickleFormat, TaxonomicClassifierDirFmt,
+                        TaxonomicClassiferTemporaryPickleDirFmt)
 plugin.register_semantic_type_to_format(
-    TaxonomicClassifier, artifact_format=TaxonomicClassifierDirFmt)
+    TaxonomicClassifier,
+    artifact_format=TaxonomicClassiferTemporaryPickleDirFmt)
